requirements.txt
black==24.10.0
colour==0.1.5
coverage==7.6.12
pathlib2==2.3.7.post1
pytest==8.3.4
svg.py==1.5.0
